requirements.txt
fastapi
uvicorn
python-multipart
pdfplumber
python-dotenv
httpx
openai
groq
jinja2
